requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
mistralai
PyMuPDF
python-docx
